requirements.txt
PyYAML==6.0.2
boto3==1.37.6
